requirements.txt
pandas
numpy
scikit-learn
dvc
pytest
flask
gunicorn
pyyaml
joblib
